# Efficient channel swapping